Core/PacketIO/DataStores/Spells: Updated packet/db2 structures to 9.0.5 & Update spell cast result enums to 9.0.5

+++ src/server/game/Entities/Player/Player.cpp
@@ -7984,7 +7984,7 @@ void Player::_ApplyItemBonuses(Item* item, uint8 slot, bool apply)
         if (statType == -1)
             continue;
 
-        int32 val = item->GetItemStatValue(i, this);
+        float val = item->GetItemStatValue(i, this);
         if (val == 0)
             continue;
